MEDProfileFigures almost done
