+++ renderer/components/input/MEDprofiles/MEDcohortFigure.jsx
@@ -1,43 +1,186 @@
+import React from "react"
 import { loadJsonPath } from "../../../utilities/fileManagementUtils"
-import React, { useEffect, useState } from "react"
-import dynamic from "next/dynamic"
-import * as d3 from "d3"
-import { Button } from "primereact/button"
-import { Col, Row } from "react-bootstrap"
-import { ToggleButton } from "primereact/togglebutton"
-import { Dropdown } from "primereact/dropdown"
 import { deepCopy } from "../../../utilities/staticFunctions"
 import { XSquare } from "react-bootstrap-icons"
 import ReactECharts from "echarts-for-react"
+import * as d3 from "d3"
+import { Col, Row } from "react-bootstrap"
+import { ToggleButton } from "primereact/togglebutton"
+import { Dropdown } from "primereact/dropdown"
+import { Button } from "primereact/button"
+import { MultiSelect } from "primereact/multiselect"
 
-/**
- *
- * @param {jsonFilePath} jsonFilePath Path to the file containing a MEDcohort as JSON data
- *
- * @returns {JSX.Element} a page
- *
- * @description
- * This component is part of the MEDprofilesViewer page and is a plotly.js plot representing
- * a MEDcohort with interactive options.
- *
- */
-const MEDcohortFigure = ({ jsonFilePath, classes, setClasses }) => {
-  const [jsonData, setJsonData] = useState(null)
-  const [plotData, setPlotData] = useState([])
-  const [selectedClass, setSelectedClass] = useState() // list of selected classes in the dropdown menu
-  const [relativeTime, setRelativeTime] = useState(null) // relative time for the selected class [0, 1
-  const [annotations, setAnnotations] = useState([])
-  const [separateVertically, setSeparateVertically] = useState(false)
-  const [separateHorizontally, setSeparateHorizontally] = useState(false)
-  const [selectedClassToSetTimePoint, setSelectedClassToSetTimePoint] = useState(null)
-  const [shapes, setShapes] = useState([])
-  const [timePoints, setTimePoints] = useState([{ label: "1", value: 1 }])
-  const [timePoint, setTimePoint] = useState(1)
-  const [selectedData, setSelectedData] = useState([])
-  const [timePointClusters, setTimePointClusters] = useState([])
-  const [test, setTest] = useState(null)
-  const [layout, setLayout] = useState({})
-  const [echartsOptions, setEchartsOptions] = useState(null)
+class MEDcohortFigureClass extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      jsonData: this.props.jsonData,
+      plotData: [],
+      selectedClass: undefined,
+      relativeTime: null,
+      annotations: [],
+      separateVertically: false,
+      separateHorizontally: false,
+      selectedClassesToSetTimePoint: null,
+      shapes: [],
+      timePoints: [{ label: "1", value: 1 }],
+      timePoint: 1,
+      selectedData: [],
+      timePointClusters: [],
+      test: null,
+      layout: {},
+      echartsOptions: null,
+      classes: this.props.classes
+    }
+    this.chartRef = React.createRef()
+  }
+
+  // You can add lifecycle methods like componentDidMount, componentDidUpdate here
+  componentDidMount() {
+    this.setState({ jsonData: loadJsonPath(this.props.jsonFilePath) }, () => {
+      this.generateEchartsOptions()
+    })
+  }
+
+  componentDidUpdate(prevProps, prevState) {
+    if (this.chartRef.current !== null) {
+      console.log("REF", this.chartRef.current.getEchartsInstance())
+    }
+    // Respond to changes in props or state
+    if (prevProps.jsonData !== this.props.jsonData) {
+      this.setState({ jsonData: this.props.jsonData }, () => {
+        this.generateEchartsOptions()
+      })
+    } else if (prevState.jsonData !== this.state.jsonData) {
+      this.generateEchartsOptions()
+    } else if (prevState.separateHorizontally !== this.state.separateHorizontally || prevState.separateVertically !== this.state.separateVertically) {
+      this.generateEchartsOptions()
+    } else if (prevState.relativeTime !== this.state.relativeTime) {
+      this.generateEchartsOptions()
+    }
+  }
+
+  componentWillUnmount() {
+    // Clean up event listeners, cancel timeouts, etc.
+  }
+
+  returnTurboColorFromIndexInList = (index, length) => {
+    return d3.interpolateTurbo(index / length)
+  }
+
+  createRectFromTimePoint = (timePoint, length, timePointClusters, echartsOptions, name) => {
+    const findEarliestDate = (timePoint) => {
+      let earliestDate = null
+      timePointClusters[timePoint].x.forEach((x, index) => {
+        if (earliestDate === null) {
+          earliestDate = x
+        } else if (x < earliestDate) {
+          earliestDate = x
+        }
+      })
+      return earliestDate
+    }
+    const findLatestDate = (timePoint) => {
+      let latestDate = new Date(0)
+      timePointClusters[timePoint].x.forEach((x) => {
+        if (x > latestDate) {
+          latestDate = x
+        }
+      })
+      return latestDate
+    }
+
+    let earliestDate = findEarliestDate(timePoint)
+    let latestDate = findLatestDate(timePoint)
+    console.log("earliestDate", earliestDate, "latestDate", latestDate, earliestDate === latestDate)
+    if (this.state.relativeTime !== null) {
+      earliestDate = earliestDate.valueOf() / (1000 * 60 * 60 * 24)
+      latestDate = latestDate.valueOf() / (1000 * 60 * 60 * 24)
+    }
+    let rect = {
+      name: `T${name}`,
+      type: "scatter",
+
+      markArea: {
+        silent: true,
+        itemStyle: {
+          color: this.returnTurboColorFromIndexInList(timePoint, 5),
+          opacity: 0.1,
+          borderWidth: 1,
+          borderType: "dashed"
+        },
+        label: {
+          position: "bottom",
+          show: true,
+          formatter: `T${name}`
+        },
+        data: [
+          [
+            {
+              name: `T${name}`,
+              xAxis: earliestDate,
+              yAxis: -1
+            },
+            {
+              xAxis: latestDate,
+              yAxis: echartsOptions.yAxis[0].data.length
+            }
+          ]
+        ]
+      },
+      data: [
+        [earliestDate, -1],
+        [latestDate, echartsOptions.yAxis[0].data.length]
+      ]
+    }
+    return rect
+  }
+
+  handleTimePointClustersChange = (timePointClusters, echartsOptions) => {
+    console.log("timePointClusters", timePointClusters)
+    let newShapes = []
+    let length = timePointClusters.length
+    timePointClusters.forEach((cluster, index) => {
+      let rect = this.createRectFromTimePoint(index, length, timePointClusters, echartsOptions, cluster.name)
+      newShapes.push(rect)
+    })
+
+    let newTimePoints = this.updateTimePoints(timePointClusters)
+    if (newTimePoints.length === timePointClusters.length) {
+      // Get last time point
+      let lastElement = newTimePoints[newTimePoints.length - 1]
+      let lastTimePoint = lastElement.value
+
+      this.range(lastTimePoint, 1).forEach((timePoint) => {
+        if (newTimePoints.findIndex((element) => element.value === timePoint) === -1) {
+          newTimePoints.push({ label: timePoint, value: timePoint })
+        }
+      })
+      if (newTimePoints.length !== 1 + timePointClusters.length) {
+        // Add the future time point
+        newTimePoints.push({ label: lastTimePoint + 1, value: lastTimePoint + 1 })
+      }
+    }
+    newTimePoints.sort((a, b) => (a.value > b.value ? 1 : -1))
+    this.setState({ timePoints: newTimePoints, shapes: newShapes })
+  }
+
+  range(size, startAt) {
+    return [...Array(size).keys()].map((i) => i + startAt)
+  }
+
+  updateTimePoints = (timePointClusters) => {
+    let newTimePoints = new Set([1])
+    timePointClusters.forEach((cluster, index) => {
+      newTimePoints.add(cluster.name)
+    })
+    console.log("newTimePoints", newTimePoints)
+    let newTimePointsArray = []
+    newTimePoints.forEach((timePoint) => {
+      newTimePointsArray.push({ label: timePoint, value: timePoint })
+    })
+    return newTimePointsArray
+  }
 
   /**
    *
@@ -48,7 +191,7 @@ const MEDcohortFigure = ({ jsonFilePath, classes, setClasses }) => {
    * @description
    * This functions returns true if the className attribute in MEDtab object is not null.
    */
-  function isNotNull(tabObject, className) {
+  isNotNull(tabObject, className) {
     let attributes = Object.keys(tabObject[className])
     let toRet = false
     attributes.forEach((attribute) => {
@@ -59,16 +202,16 @@ const MEDcohortFigure = ({ jsonFilePath, classes, setClasses }) => {
     return toRet
   }
 
-  const getTimeZeroForClass = (className, profileIndex) => {
+  getTimeZeroForClass = (className, profileIndex) => {
     let timeZeroAttribute = null
     if (className === null) return null
-    jsonData?.list_MEDprofile[profileIndex]?.list_MEDtab?.forEach((tab) => {
+    this.state.jsonData?.list_MEDprofile[profileIndex]?.list_MEDtab?.forEach((tab) => {
       let attributes = Object.keys(tab)
       let attributeIndex = attributes.indexOf(className)
       if (attributeIndex !== -1) {
         let attribute = attributes[attributeIndex]
-        if (attribute !== "Date" && attribute !== "Time_point" && isNotNull(tab, attribute)) {
-          if (attribute === relativeTime && timeZeroAttribute === null && tab.Date !== null) {
+        if (attribute !== "Date" && attribute !== "Time_point" && this.isNotNull(tab, attribute)) {
+          if (attribute === this.state.relativeTime && timeZeroAttribute === null && tab.Date !== null) {
             return (timeZeroAttribute = tab.Date)
           }
         }
@@ -79,84 +222,29 @@ const MEDcohortFigure = ({ jsonFilePath, classes, setClasses }) => {
     return timeZeroAttribute
   }
 
-  // Format the JSON data in order to display it in the graph
-  // const formatData = () => {
-  //   let formattedData = []
-  //   let newClasses = new Set()
-  //   let timeZeroAttribute = null
-  //   let newTimePointsClusters = []
-  //   jsonData?.list_MEDprofile?.forEach((profile, index) => {
-  //     const color = d3.interpolateTurbo(jsonData.list_MEDprofile.indexOf(profile) / jsonData.list_MEDprofile.length)
-  //     let profileData = { x: [], y: [], mode: "markers", type: "scatter", marker: { color: color }, text: [], name: profile.PatientID, customdata: [] }
-  //     let profileRandomTime = index
-  //     let profilAttributeTimeZero = getTimeZeroForClass(relativeTime, index)
-  //     profile?.list_MEDtab?.forEach((tab) => {
-  //       let attributes = Object.keys(tab)
-  //       attributes.forEach((attribute) => {
-  //         newClasses.add(attribute)
-  //         if (attribute !== "Date") {
-  //           if (attribute === relativeTime && timeZeroAttribute === null) {
-  //             timeZeroAttribute = 0
-  //           }
-  //           let newDate = new Date(tab.Date)
-  //           if (profilAttributeTimeZero !== null) {
-  //             newDate = (new Date(tab.Date) - new Date(profilAttributeTimeZero)) / (1000 * 60 * 60 * 24 * 365)
-  //           }
-  //           if (separateHorizontally) {
-  //             newDate.setHours(newDate.getHours() + profileRandomTime)
-  //           }
-  //           if (attribute !== "Time_point" && isNotNull(tab, attribute)) {
-  //             profileData.x.push(newDate)
-  //             if (separateVertically) {
-  //               profileData.y.push(attribute + profileRandomTime)
-  //             } else {
-  //               profileData.y.push(attribute)
-  //             }
-  //             profileData.text.push(`${attribute}`)
-  //             profileData.customdata.push(tab[attribute])
-  //           } else if (attribute === "Time_point") {
-  //             let timePoint = tab[attribute]
-  //             if (timePoint === null) return
-  //             if (newTimePointsClusters[timePoint] === undefined || newTimePointsClusters[timePoint] === null) {
-  //               newTimePointsClusters[timePoint] = { x: [], y: [], mode: "lines", type: "scatter", marker: { color: color }, text: [], name: timePoint, customdata: [], fill: "toself" }
-  //             }
-  //             newTimePointsClusters[timePoint].x.push(newDate)
-  //             if (separateVertically) {
-  //               newTimePointsClusters[timePoint].y.push(attribute + profileRandomTime)
-  //             } else {
-  //               newTimePointsClusters[timePoint].y.push(attribute)
-  //             }
-  //           }
-  //         }
-  //       })
-  //     })
+  handleSetTimePointByClass = () => {
+    console.log("selectedClassesToSetTimePoint", this.state.selectedClassesToSetTimePoint)
+    let newJsonData = { ...this.state.jsonData }
+    newJsonData.list_MEDprofile.forEach((profile) => {
+      profile.list_MEDtab.forEach((tab) => {
+        let attributes = Object.keys(tab)
+        attributes.forEach((attribute) => {
+          if (attribute !== "Date" && attribute !== "Time_point" && this.isNotNull(tab, attribute)) {
+            if (this.state.selectedClassesToSetTimePoint.includes(attribute)) {
+              if (tab.Time_point === null) {
+                tab.Time_point = [this.state.timePoint]
+              } else {
+                tab.Time_point.push(this.state.timePoint)
+              }
+            }
+          }
+        })
+      })
+    })
+    this.setState({ jsonData: newJsonData })
+  }
 
-  //     formattedData.push(profileData)
-  //   })
-  //   let correctedTimePointClusters = []
-  //   Object.keys(newTimePointsClusters).forEach((key) => {
-  //     correctedTimePointClusters.push(newTimePointsClusters[key])
-  //   })
-  //   setPlotData(formattedData)
-  //   setClasses(newClasses)
-  //   setTimePointClusters(correctedTimePointClusters)
-  // }
-
-  // Called at initialization in order to load the JSON data
-  useEffect(() => {
-    setJsonData(loadJsonPath(jsonFilePath))
-  }, [])
-
-  // Called while JSON data is loaded in order to call the formatData() function
-  useEffect(() => {
-    if (jsonData) {
-      // formatData()
-      // formatData()
-      generateEchartsOptions()
-    }
-  }, [jsonData])
-
-  const generateEchartsOptions = () => {
+  generateEchartsOptions = () => {
     let newEchartsOption = {
       title: {
         text: "MEDcohort"
@@ -172,7 +260,7 @@ const MEDcohortFigure = ({ jsonFilePath, classes, setClasses }) => {
       },
       xAxis: [
         {
-          type: (relativeTime !== null && "value") || "time"
+          type: (this.state.relativeTime !== null && "value") || "time"
         }
       ],
       yAxis: [
@@ -182,10 +270,10 @@ const MEDcohortFigure = ({ jsonFilePath, classes, setClasses }) => {
         }
       ],
       brush: {
-        brushLink: "all",
-        toolbox: ["rect", "polygon", "clear"],
+        toolbox: ["rect", "polygon", "lineX", "lineY", "clear"],
         seriesIndex: "all",
-        xAxisIndex: 0,
+        xAxisIndex: "all",
+        yAxisIndex: "all",
         inBrush: {
           opacity: 1
         },
@@ -227,46 +315,46 @@ const MEDcohortFigure = ({ jsonFilePath, classes, setClasses }) => {
       }
       // visualMap: {}
     }
-
+    // console.log("this.state.jsonData", this.state.jsonData)
     let patientNames = new Set()
     let innerYClasses = new Set()
     let newClasses = new Set()
     let timeZeroAttribute = 0
     let newTimePointsClusters = []
-    let numberOfPatients = jsonData?.list_MEDprofile?.length
-    jsonData?.list_MEDprofile?.forEach((profile, index) => {
+    let numberOfPatients = this.state.jsonData?.list_MEDprofile?.length
+    this.state.jsonData?.list_MEDprofile?.forEach((profile, index) => {
       if (profile.PatientID !== "32379" && profile.PatientID !== "25881" && profile.PatientID !== "21690" && profile.PatientID !== "18089") {
-        const color = d3.interpolateTurbo(jsonData.list_MEDprofile.indexOf(profile) / jsonData.list_MEDprofile.length)
+        const color = d3.interpolateTurbo(this.state.jsonData.list_MEDprofile.indexOf(profile) / this.state.jsonData.list_MEDprofile.length)
         patientNames.add(profile.PatientID)
         let profileSerie = { type: "scatter", data: [], name: profile.PatientID, itemStyle: { color: color }, symbolSize: 5, emphasis: { focus: "series" }, selectMode: "multiple" }
         let profileRandomTime = index
-        let profilAttributeTimeZero = getTimeZeroForClass(relativeTime, index)
+        let profilAttributeTimeZero = this.getTimeZeroForClass(this.state.relativeTime, index)
         profile?.list_MEDtab?.forEach((tab) => {
           let attributes = Object.keys(tab)
           attributes.forEach((attribute) => {
             newClasses.add(attribute)
             if (attribute !== "Date") {
-              if (attribute === relativeTime && timeZeroAttribute === null) {
+              if (attribute === this.state.relativeTime && timeZeroAttribute === null) {
                 timeZeroAttribute = tab.Date
               }
               let newDate = new Date(tab.Date)
               if (profilAttributeTimeZero !== null) {
                 newDate = new Date(new Date(tab.Date) - new Date(profilAttributeTimeZero))
               }
-              if (separateHorizontally) {
+              if (this.state.separateHorizontally) {
                 newDate = Date.parse(newDate + profileRandomTime)
               }
               let x, y
-              if (attribute !== "Time_point" && isNotNull(tab, attribute)) {
-                if (relativeTime !== null) {
+              if (attribute !== "Time_point" && this.isNotNull(tab, attribute)) {
+                if (this.state.relativeTime !== null) {
                   x = newDate.valueOf() / (1000 * 60 * 60 * 24)
-                  if (separateHorizontally) {
+                  if (this.state.separateHorizontally) {
                     x = x + profileRandomTime / (numberOfPatients * 2)
                   }
                 } else {
                   x = newDate
                 }
-                if (separateVertically) {
+                if (this.state.separateVertically) {
                   y = attribute + profileRandomTime
                   innerYClasses.add(attribute + profileRandomTime)
                 } else {
@@ -275,25 +363,25 @@ const MEDcohortFigure = ({ jsonFilePath, classes, setClasses }) => {
                 }
                 profileSerie.data.push([x, y])
               } else if (attribute === "Time_point") {
-                let timePoint = tab[attribute]
-                if (timePoint === null) return
-                if (newTimePointsClusters[timePoint] === undefined || newTimePointsClusters[timePoint] === null) {
-                  newTimePointsClusters[timePoint] = { x: [], y: [], mode: "lines", type: "scatter", marker: { color: color }, text: [], name: timePoint, customdata: [], fill: "toself" }
-                }
-                newTimePointsClusters[timePoint].x.push(newDate)
-                if (separateVertically) {
-                  newTimePointsClusters[timePoint].y.push(attribute + profileRandomTime)
-                } else {
-                  newTimePointsClusters[timePoint].y.push(attribute)
-                }
+                let timePoints = tab[attribute]
+                if (timePoints === null) return
+                timePoints.forEach((timePoint) => {
+                  if (newTimePointsClusters[timePoint] === undefined || newTimePointsClusters[timePoint] === null) {
+                    newTimePointsClusters[timePoint] = { x: [], y: [], mode: "lines", type: "scatter", marker: { color: color }, text: [], name: timePoint, customdata: [], fill: "toself" }
+                  }
+                  newTimePointsClusters[timePoint].x.push(newDate)
+                  if (this.state.separateVertically) {
+                    newTimePointsClusters[timePoint].y.push(attribute + profileRandomTime)
+                  } else {
+                    newTimePointsClusters[timePoint].y.push(attribute)
+                  }
+                })
               }
             }
           })
         })
         newEchartsOption.series.push(profileSerie)
       }
-
-      // formattedData.push(profileData)
     })
     newEchartsOption.yAxis[0].data = [...innerYClasses]
     newEchartsOption.legend.data = [...patientNames]
@@ -301,97 +389,51 @@ const MEDcohortFigure = ({ jsonFilePath, classes, setClasses }) => {
     Object.keys(newTimePointsClusters).forEach((key) => {
       correctedTimePointClusters.push(newTimePointsClusters[key])
     })
-    setEchartsOptions(newEchartsOption)
-    setClasses(newClasses)
-    setTimePointClusters(correctedTimePointClusters)
+    this.handleTimePointClustersChange(correctedTimePointClusters, newEchartsOption)
+
+    this.setState({ echartsOptions: newEchartsOption })
+    this.setState({ timePointClusters: correctedTimePointClusters })
+    this.setState({ classes: newClasses })
   }
 
-  useEffect(() => {
-    console.log("echartsOptions", echartsOptions)
-  }, [echartsOptions])
-
-  // If the relativeTime is changed, we update the figure
-  useEffect(() => {
-    console.log("relativeTime", relativeTime)
-    // formatData()
-    generateEchartsOptions()
-  }, [relativeTime])
-
-  useEffect(() => {
-    console.log("plotData", plotData, jsonData)
-  }, [plotData])
-
-  const getFindPatientFunction = (patientId) => {
-    return (patient) => {
-      return patient.PatientID === patientId
-    }
-  }
-
-  const handleClick = (data) => {
-    var point = data.points[0]
-    console.log("point", point)
-    let customdata = point.customdata
-    let textToShow = "<div>"
-    console.log("customdata", customdata)
-
-    textToShow += "<b>Patient</b> " + point.data.name + "<br>"
-    textToShow += "<b>Time</b> " + point.x + "<br>"
-    textToShow += "<b>Class</b> " + point.y + "<br>"
-    Object.entries(customdata).forEach((data) => {
-      textToShow += "<b>" + data[0] + "</b> " + data[1] + "<br>"
+  removeTimePointFromJsonData = (timePoint) => {
+    let newJsonData = { ...this.state.jsonData }
+    console.log("timepoint to remove", timePoint)
+    newJsonData.list_MEDprofile.forEach((profile) => {
+      profile.list_MEDtab.forEach((tab) => {
+        if (tab.Time_point === null || tab.Time_point === undefined || tab.Time_point.length === 0) return
+        console.log("REMOVE TAB", tab.Time_point, timePoint, tab.Time_point.includes(parseInt(timePoint)))
+        if (tab.Time_point.includes(parseInt(timePoint))) {
+          let timePointIndex = tab.Time_point.findIndex((timePointElement) => timePointElement === timePoint)
+          tab.Time_point.splice(timePointIndex, 1)
+        }
+      })
     })
-    textToShow += "</div>"
-    let newAnnotation = {
-      x: point.x,
-      y: point.y,
-      arrowhead: 6,
-      ax: 0,
-      ay: -80,
-      bgcolor: "rgba(255, 255, 255, 0.9)",
-      arrowcolor: point.fullData.marker.color,
-      font: { size: 11 },
-      bordercolor: point.fullData.marker.color,
-      borderwidth: 3,
-      borderpad: 4,
-      height: 120,
-      width: 200,
-
-      text: textToShow
-    }
-
-    console.log(point.pointNumber)
-    let newAnnotations = [...annotations]
-    let toDelete = false
-    newAnnotations.forEach((annotation) => {
-      if (annotation.x === newAnnotation.x && annotation.y === newAnnotation.y) {
-        newAnnotations.splice(newAnnotations.indexOf(annotation), 1)
-        toDelete = true
+    let newShapes = [...this.state.shapes]
+    console.log("newJsonData", newJsonData)
+    this.state.shapes.forEach((shape) => {
+      if (shape.name === `T${timePoint}`) {
+        let shapeIndex = newShapes.findIndex((shapeElement) => shapeElement.name === `T${timePoint}`)
+        newShapes.splice(shapeIndex, 1)
       }
     })
-    if (toDelete) {
-      setAnnotations(newAnnotations)
-      return
-    }
+    console.log("newShapes", newShapes, this.state.shapes, this.state.shapes.length, newShapes.length)
+    this.setState({ shapes: newShapes, jsonData: newJsonData }, () => {
+      this.generateEchartsOptions()
+      let newEchartsOptions = { ...this.state.echartsOptions }
+      newEchartsOptions.series = [...this.state.echartsOptions.series, ...newShapes]
 
-    newAnnotations.push(newAnnotation)
-    setAnnotations(newAnnotations)
-    // delete instead if clicked twic
+      this.chartRef.current.getEchartsInstance().setOption(newEchartsOptions, { notMerge: true })
+      console.log("newEchartsOptions", newEchartsOptions, this.chartRef.current.getEchartsInstance().getOption())
+    })
   }
 
-  useEffect(() => {
-    console.log("separate", separateVertically, separateHorizontally)
-    // formatData()
-    generateEchartsOptions()
-  }, [separateVertically, separateHorizontally])
-
-  useEffect(() => {
-    console.log("annotations", annotations)
-  }, [annotations])
-
-  const getClassesOptions = () => {
-    if (classes.size === 0) return []
+  getClassesOptions = () => {
+    let thisClasses = this.state.classes
+    if (thisClasses === null) return []
+    if (thisClasses.size === 0) return []
     let classesArray = []
-    classes.forEach((className) => {
+    thisClasses.forEach((className) => {
       if (className !== "Date" && className !== "Time_point") {
         classesArray.push({ label: className, value: className })
       }
@@ -400,399 +442,223 @@ const MEDcohortFigure = ({ jsonFilePath, classes, setClasses }) => {
     return classesArray
   }
 
-  const handleSetTimePointByClass = () => {
-    console.log("selectedClassToSetTimePoint", selectedClassToSetTimePoint)
-    let newJsonData = { ...jsonData }
-    newJsonData.list_MEDprofile.forEach((profile) => {
-      profile.list_MEDtab.forEach((tab) => {
+  getFindPatientFunction = (patientId) => {
+    return (patient) => {
+      return patient.PatientID === patientId
+    }
+  }
+
+  handleSetTimePoint = () => {
+    console.log("SETTINT TIMEPOINT", this.state.timePoint)
+    console.log("selectedData", this.state.selectedData, this.state.jsonData.list_MEDprofile)
+    let newJsonData = { ...this.state.jsonData }
+    /*
+    selectedData: [
+      {
+        seriesId: '\x001693\x000',
+        seriesName: '1693',
+        seriesIndex: 0, 
+        dataIndex: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9],
+      }, 
+      ...
+    ]
+    */
+
+    this.state.selectedData.forEach((data, seriesIndex) => {
+      let seriesName = data.seriesName
+      let profileIndex = newJsonData.list_MEDprofile.findIndex(this.getFindPatientFunction(seriesName))
+      let profile = newJsonData.list_MEDprofile[profileIndex]
+      let selectedPoints = data.dataIndex
+      console.log("selectedPoints", selectedPoints, profile, this.state.echartsOptions.series[seriesIndex])
+      if (profile === undefined) return
+      let correspondingData = this.state.echartsOptions.series[seriesIndex].data
+      let onlySelectedData = []
+      correspondingData.forEach((dataPoint, index) => {
+        if (selectedPoints.includes(index)) {
+          onlySelectedData.push(dataPoint)
+        }
+      })
+      console.log("onlySelectedData", onlySelectedData)
+      let patientGlobalIndex = 0
+      profile.list_MEDtab.forEach((tab, index) => {
         let attributes = Object.keys(tab)
         attributes.forEach((attribute) => {
-          if (attribute !== "Date" && attribute !== "Time_point" && isNotNull(tab, attribute)) {
-            if (attribute === selectedClassToSetTimePoint) {
-              tab.Time_point = timePoint
+          if (attribute !== "Date" && attribute !== "Time_point" && this.isNotNull(tab, attribute)) {
+            if (selectedPoints.includes(patientGlobalIndex)) {
+              console.log("patientGlobalIndex", patientGlobalIndex, tab, attribute)
+              if (tab.Time_point === null) {
+                tab.Time_point = [this.state.timePoint]
+              } else {
+                if (!tab.Time_point.includes(this.state.timePoint)) {
+                  tab.Time_point.push(this.state.timePoint)
+                }
+              }
             }
+            patientGlobalIndex += 1
           }
         })
       })
     })
-    setJsonData(newJsonData)
+    // profile.list_MEDtab.forEach((tab, index) => {
+
+    // selectedPoints.forEach((pointIndex) => {
+    //   let tab = profile.list_MEDtab[pointIndex]
+    //   if (tab !== undefined) {
+    //     console.log("pointIndex", pointIndex, profile, tab)
+    //     if (tab.Time_point === null) {
+    //       tab.Time_point = [this.state.timePoint]
+    //     } else {
+    //       tab.Time_point.push(this.state.timePoint)
+    //     }
+    //   }
+    // })
+
+    //   console.log("data", data, newJsonData.list_MEDprofile)
+    //   let profileIndex = newJsonData.list_MEDprofile.findIndex(this.getFindPatientFunction(data.name))
+
+    //   // Go through all the profiles and set the time point to the selected value for the selected data
+    //   let profile = newJsonData.list_MEDprofile[profileIndex]
+    //   let selectedPoints = data.selectedpoints
+    //   selectedPoints.forEach((pointIndex) => {
+    //     let tab = profile.list_MEDtab[pointIndex]
+    //     console.log("pointIndex", pointIndex, profile, tab)
+    //     if (tab.Time_point === null) {
+    //       tab.Time_point = [this.state.timePoint]
+    //     } else {
+    //       tab.Time_point.push(this.state.timePoint)
+    //     }
+    //   })
+    // })
+    this.setState({ jsonData: newJsonData })
   }
 
-  const handleSetTimePoint = () => {
-    console.log("timePoint", timePoint)
-    console.log("selectedData", selectedData)
-    let newJsonData = { ...jsonData }
-    selectedData.forEach((data) => {
-      console.log("data", data, newJsonData.list_MEDprofile)
-      let profileIndex = newJsonData.list_MEDprofile.findIndex(getFindPatientFunction(data.name))
-
-      // Go through all the profiles and set the time point to the selected value for the selected data
-      let profile = newJsonData.list_MEDprofile[profileIndex]
-      let selectedPoints = data.selectedpoints
-      selectedPoints.forEach((pointIndex) => {
-        let tab = profile.list_MEDtab[pointIndex]
-        console.log("pointIndex", pointIndex, profile, tab)
-        tab.Time_point = timePoint
-      })
-    })
-    setJsonData(newJsonData)
+  handleSelectData = (data) => {
+    // console.log("data", data.batch["0"].selected)
+    this.setState({ selectedData: data.batch["0"].selected })
   }
 
-  useEffect(() => {
-    console.log("selectedData", selectedData)
-  }, [selectedData])
-
-  useEffect(() => {
+  render() {
+    // Destructure state and props for easier access
+    const { classes, setClasses, jsonFilePath } = this.props
+    const { jsonData, plotData, selectedClass, relativeTime, annotations, separateVertically, separateHorizontally, selectedClassesToSetTimePoint, shapes, timePoints, timePoint, selectedData, timePointClusters, test, layout, echartsOptions } = this.state
+    let newEchartsOption = { ...echartsOptions }
     if (echartsOptions !== null) {
-      let newEchartsOptions = { ...echartsOptions }
-      shapes.forEach((shape) => {
-        newEchartsOptions.series.push(shape)
-      })
-      setEchartsOptions(newEchartsOptions)
+      newEchartsOption.series = [...echartsOptions.series, ...shapes]
     }
-  }, [shapes])
+    // Render your component's JSX here
+    return (
+      <>
+        <Row style={{ width: "100%", justifyContent: "center" }}>
+          <Col lg={8} className="center">
+            <div className="MEDcohort-figure" style={{ display: "flex", flexDirection: "column", boxShadow: "2px 2px 4px rgba(0, 0, 0, 0.25)" }}>
+              {echartsOptions && <ReactECharts ref={this.chartRef} option={newEchartsOption} onEvents={{ brushselected: this.handleSelectData }} style={{ width: "100%", height: "100%" }} lazyUpdate={false} />}
+            </div>
+          </Col>
 
-  useEffect(() => {
-    console.log("timePointClusters", timePointClusters)
-    let newShapes = []
-    let length = timePointClusters.length
-    timePointClusters.forEach((cluster, index) => {
-      let rect = createRectFromTimePoint(index, length)
-      newShapes.push(rect)
-    })
-    setShapes(newShapes)
-
-    let newTimePoints = updateTimePoints()
-    if (newTimePoints.length === timePointClusters.length) {
-      // Get last time point
-      let lastElement = newTimePoints[newTimePoints.length - 1]
-      let lastTimePoint = lastElement.value
-
-      range(lastTimePoint, 1).forEach((timePoint) => {
-        if (newTimePoints.findIndex((element) => element.value === timePoint) === -1) {
-          newTimePoints.push({ label: timePoint, value: timePoint })
-        }
-      })
-      if (newTimePoints.length !== 1 + timePointClusters.length) {
-        // Add the future time point
-        newTimePoints.push({ label: lastTimePoint + 1, value: lastTimePoint + 1 })
-      }
-    }
-    newTimePoints.sort((a, b) => (a.value > b.value ? 1 : -1))
-    setTimePoints(newTimePoints)
-  }, [timePointClusters])
-
-  function range(size, startAt) {
-    return [...Array(size).keys()].map((i) => i + startAt)
-  }
-
-  const updateTimePoints = () => {
-    let newTimePoints = new Set([1])
-    timePointClusters.forEach((cluster, index) => {
-      newTimePoints.add(cluster.name)
-    })
-    console.log("newTimePoints", newTimePoints)
-    let newTimePointsArray = []
-    newTimePoints.forEach((timePoint) => {
-      newTimePointsArray.push({ label: timePoint, value: timePoint })
-    })
-    return newTimePointsArray
-  }
-
-  const createRectFromTimePoint = (timePoint, length) => {
-    const findEarliestDate = (timePoint) => {
-      let earliestDate = null
-      timePointClusters[timePoint].x.forEach((x, index) => {
-        if (earliestDate === null) {
-          earliestDate = x
-        } else if (x < earliestDate) {
-          earliestDate = x
-        }
-      })
-      return earliestDate
-    }
-    const findLatestDate = (timePoint) => {
-      let latestDate = new Date(0)
-      timePointClusters[timePoint].x.forEach((x, index) => {
-        if (x > latestDate) {
-          latestDate = x
-        }
-      })
-      return latestDate
-    }
-    let earliestDate = findEarliestDate(timePoint)
-    let latestDate = findLatestDate(timePoint)
-    console.log("earliestDate", earliestDate, "latestDate", latestDate, earliestDate === latestDate)
-    // if (earliestDate.toTimeString() === latestDate.toTimeString()) {
-    //   // earliestDate.setHours(earliestDate.getHours() - 1)
-    //   // latestDate.setHours(latestDate.getHours() + 1)
-    // }
-    if (relativeTime !== null) {
-      earliestDate = earliestDate.valueOf() / (1000 * 60 * 60 * 24)
-      latestDate = latestDate.valueOf() / (1000 * 60 * 60 * 24)
-    }
-    let rect = {
-      name: `T${timePoint + 1}`,
-      type: "scatter",
-
-      markArea: {
-        silent: true,
-        itemStyle: {
-          color: returnTurboColorFromIndexInList(timePoint, 5),
-          opacity: 0.1,
-          borderWidth: 1,
-          borderType: "dashed"
-        },
-        label: {
-          position: "bottom",
-          show: true,
-          formatter: `T${timePoint + 1}`
-        },
-        data: [
-          [
-            {
-              name: `T${timePoint + 1}`,
-              xAxis: earliestDate,
-              yAxis: -1
-            },
-            {
-              xAxis: latestDate,
-              yAxis: echartsOptions.yAxis[0].data.length
-            }
-          ]
-        ]
-      },
-      data: [
-        [earliestDate, -1],
-        [latestDate, echartsOptions.yAxis[0].data.length]
-      ]
-    }
-    return rect
-  }
-
-  const darkenColorFromTurbo = (color) => {
-    let rgb = d3.rgb(color)
-    let hsl = d3.hsl(rgb)
-    hsl.l = hsl.l - 0.2
-    return hsl.toString()
-  }
-
-  const addTransparencyToColor = (color, transparency) => {
-    let rgb = d3.rgb(color)
-    let hsl = d3.hsl(rgb)
-    hsl.opacity = transparency
-    return hsl.toString()
-  }
-
-  const returnTurboColorFromIndexInList = (index, length) => {
-    return d3.interpolateTurbo(index / length)
-  }
-
-  const removeTimePointFromJsonData = (timePoint) => {
-    let newJsonData = { ...jsonData }
-    console.log("timepoint to remove", timePoint)
-    newJsonData.list_MEDprofile.forEach((profile) => {
-      profile.list_MEDtab.forEach((tab) => {
-        if (tab.Time_point === timePoint) {
-          tab.Time_point = null
-        }
-      })
-    })
-    setJsonData(newJsonData)
-  }
-
-  useEffect(() => {
-    console.log("Layout", layout)
-  }, [layout])
-
-  // const handleSelect = (param, echartSelected) => {
-  //   console.log("selected", echartSelected, param)
-  //   // let batch = param.batch
-  //   if (param.areas === null) return
-  //   else if (param.areas !== undefined) {
-  //     // let areas = batch["0"].areas
-  //     // let newSelectedData = []
-  //     // areas.forEach((area) => {
-  //     //   newSelectedData.push({ type: "lineX", range: area.coordRanges })
-  //     // })
-  //     // setSelectedData(param.areas)
-  //   }
-  // }
-  // const handleSelect = (param, echartSelected) => {
-  //   console.log("selected", echartSelected, param)
-  //   let batch = param.batch
-  //   if (batch === null) return
-  //   else if (batch["0"].areas !== undefined) {
-  //     let areas = batch["0"].areas
-  //     let newSelectedData = []
-  //     areas.forEach((area) => {
-  //       newSelectedData.push({ type: "lineX", range: area.coordRanges })
-  //     })
-  //     // setSelectedData(newSelectedData)
-  //   }
-  // }
-
-  const ref = React.useRef(null)
-  // console.log("echartsOptions", ref.current.getEchartsInstance())
-  return (
-    <>
-      <Row style={{ width: "100%", justifyContent: "center" }}>
-        <Col lg={8} className="center">
-          <div className="MEDcohort-figure" style={{ display: "flex", flexDirection: "column", boxShadow: "2px 2px 4px rgba(0, 0, 0, 0.25)" }}>
-            {echartsOptions && <ReactECharts ref={ref} option={echartsOptions} onEvents={{}} style={{ width: "100%", height: "100%" }} lazyUpdate={true} />}
-
-            {/* <Plot
-              data={plotData}
-              onClick={(data) => {
-                console.log("data", data, plotData, jsonData)
-                handleClick(data)
-              }}
-              onClickAnnotation={(data) => {
-                console.log("ANNOTATION", data)
-              }}
-              layout={{
-                // Add sliders with multiple timestamps
-                sliders: [
-                  // Add a slider for the relative time
-                ],
-                shapes: shapes,
-                selections: [],
-                selectdirection: "h",
-                autosize: true,
-                title: "MEDcohort",
-                showlegend: true,
-                legend: {
-                  title: {
-                    text: "<b>Patients</b>"
-                  }
-                },
-                xaxis: {
-                  title: "<b>Date</b>",
-                  type: "date"
-                  // rangeslider: {}
-                },
-                yaxis: {
-                  title: "<b>Classes</b>",
-                  type: "category",
-                  automargin: true
-                },
-                annotations: annotations
-              }}
-              useResizeHandler={true}
-              onUpdate={(figure) => {
-                console.log("UPDATE LAYOUT", figure)
-                console.log("Selections", figure.layout.selections)
-              }}
-              onInitialized={(figure) => {
-                console.log("INITIALIZED", figure)
-              }}
-              style={{ width: "100%", height: "100%" }}
-            /> */}
-          </div>
-        </Col>
-
-        <Col lg={4} style={{ display: "flex", flexDirection: "column", justifyContent: "space-evenly" }}>
-          <Row className="justify-content-md-center medprofile-buttons" style={{ display: "flex", flexDirection: "row", alignContent: "center", alignItems: "center", width: "100%", justifyContent: "center", boxShadow: "2px 2px 4px rgba(0, 0, 0, 0.25)", padding: "1rem", borderRadius: "1rem" }}>
-            <Col xxl="6" style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginBottom: "1rem" }}>
-              <ToggleButton className="separate-toggle-button" checked={separateHorizontally} onChange={(e) => setSeparateHorizontally(e.value)} onLabel="Overlap horizontally" offLabel="Separate horizontally" onIcon="pi pi-check" offIcon="pi pi-times" />
-            </Col>
-            <Col xxl="6" style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginBottom: "1rem" }}>
-              <ToggleButton className="separate-toggle-button" checked={separateVertically} onChange={(e) => setSeparateVertically(e.value)} onLabel="Overlap vertically" offLabel="Separate vertically" onIcon="pi pi-check" offIcon="pi pi-times" />
-            </Col>
-            <Col style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginBottom: "1rem" }}>
-              <Button size="small" label="Clear annotations" onClick={() => setAnnotations([])} />
-            </Col>
-            <Col style={{ display: "flex", flexDirection: "column", justifyContent: "center" }}>
-              <label htmlFor="dd-city">Select the class for relative time</label>
-              <Col style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginTop: "0rem" }}>
-                <div style={{ width: "100%" }} className="p-inputgroup ">
-                  <Dropdown style={{ width: "100%" }} value={selectedClass} options={getClassesOptions()} onChange={(e) => setSelectedClass(e.value)} />
-                  <ToggleButton
-                    className={`relative-time-toggle-button ${relativeTime !== null ? "p-button-success" : "p-button-info"}`}
-                    checked={relativeTime === null}
-                    onLabel="Set"
-                    offLabel="Unset"
-                    onChange={(e) => {
-                      if (relativeTime === null) {
-                        setRelativeTime(selectedClass)
-                      } else {
-                        setRelativeTime(null)
-                      }
-                    }}
-                    style={{ borderRadius: "0 4px 4px 0" }}
-                  />
-                </div>
+          <Col lg={4} style={{ display: "flex", flexDirection: "column", justifyContent: "space-evenly" }}>
+            <Row className="justify-content-md-center medprofile-buttons" style={{ display: "flex", flexDirection: "row", alignContent: "center", alignItems: "center", width: "100%", justifyContent: "center", boxShadow: "2px 2px 4px rgba(0, 0, 0, 0.25)", padding: "1rem", borderRadius: "1rem" }}>
+              <Col xxl="6" style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginBottom: "1rem" }}>
+                <ToggleButton className="separate-toggle-button" checked={separateHorizontally} onChange={(e) => this.setState({ separateHorizontally: e.value })} onLabel="Overlap horizontally" offLabel="Separate horizontally" onIcon="pi pi-check" offIcon="pi pi-times" />
               </Col>
-            </Col>
-            <Col style={{ display: "flex", flexDirection: "column", justifyContent: "center", marginTop: "1rem" }}>
-              <label htmlFor="dd-city">Set time points to selected data points</label>
-              <Col style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginTop: "0rem" }}>
-                <div style={{ width: "100%" }} className="p-inputgroup ">
-                  <Dropdown style={{ width: "100%" }} value={timePoint} options={timePoints} onChange={(e) => setTimePoint(e.value)} />
-                  <Button className="separate-toggle-button" style={{ borderRadius: "0 4px 4px 0", width: "5rem", padding: "0rem" }} onClick={handleSetTimePoint} label={`Set T${timePoint}`} />
-                </div>
+              <Col xxl="6" style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginBottom: "1rem" }}>
+                <ToggleButton className="separate-toggle-button" checked={separateVertically} onChange={(e) => this.setState({ separateVertically: e.value })} onLabel="Overlap vertically" offLabel="Separate vertically" onIcon="pi pi-check" offIcon="pi pi-times" />
               </Col>
-            </Col>
-            <Col style={{ display: "flex", flexDirection: "column", justifyContent: "center", marginTop: "1rem" }}>
-              <label htmlFor="dd-city">Set time points by classes</label>
-              <Col style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginTop: "0rem" }}>
-                <div style={{ width: "100%" }} className="p-inputgroup ">
-                  <Dropdown style={{ width: "100%" }} value={selectedClassToSetTimePoint} options={getClassesOptions()} onChange={(e) => setSelectedClassToSetTimePoint(e.value)} />
-                  <Button className="separate-toggle-button" style={{ borderRadius: "0 4px 4px 0", width: "5rem", padding: "0rem" }} onClick={handleSetTimePointByClass} label={`Set T${timePoint}`} />
-                </div>
+              <Col style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginBottom: "1rem" }}>
+                <Button size="small" label="Clear annotations" onClick={() => this.setState({ annotations: [] })} />
               </Col>
-            </Col>
-            <Col style={{ display: "flex", flexDirection: "column", justifyContent: "center", marginTop: "1rem", alignItems: "flex-start" }}>
-              {timePointClusters.length !== 0 && (
-                <>
-                  <label htmlFor="">
-                    <b>
-                      <u>Time Points associated data</u>
-                    </b>
-                  </label>
-                  {timePointClusters.map((cluster, index) => {
-                    console.log("cluster", cluster)
-                    return (
-                      <>
-                        <div style={{ display: "flex", flexDirection: "row", alignContent: "center", alignItems: "flex-start", justifyContent: "center" }}>
-                          <h6 style={{ margin: "0" }}>{`T${cluster.name}`}</h6>
-                          &nbsp;
-                          <p style={{ margin: "0", marginLeft: "0.5rem" }}>
-                            {" "}
-                            {`Number of data points: `}
-                            <b> {`${cluster.x.length}`}</b>
-                          </p>
-                          <a
-                            value={cluster.name}
-                            style={{ margin: "0", marginLeft: "0.5rem", cursor: "pointer" }}
-                            onClick={(e) => {
-                              let newTimePointsClusters = deepCopy(timePointClusters)
-                              newTimePointsClusters.splice(index, 1)
-                              let newTimePoints = deepCopy(timePoints)
-                              let indexOfTimePoint = newTimePoints.findIndex((timePoint) => timePoint.value === cluster.name)
-                              console.log("indexOfTimePoint", indexOfTimePoint)
-                              if (indexOfTimePoint !== -1 && cluster.name !== 1) {
-                                newTimePoints.splice(indexOfTimePoint, 1)
-                              }
-                              setTimePoints(newTimePoints)
-                              setTimePointClusters(newTimePointsClusters)
-                              removeTimePointFromJsonData(cluster.name)
-                            }}
-                          >
-                            <XSquare size={20} />
-                          </a>
-                        </div>
-                      </>
-                    )
-                  })}
-                </>
-              )}
-            </Col>
-          </Row>
-        </Col>
-      </Row>
-    </>
-  )
+              <Col style={{ display: "flex", flexDirection: "column", justifyContent: "center" }}>
+                <label htmlFor="dd-city">Select the class for relative time</label>
+                <Col style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginTop: "0rem" }}>
+                  <div style={{ width: "100%" }} className="p-inputgroup ">
+                    <Dropdown style={{ width: "100%" }} value={selectedClass} options={this.getClassesOptions()} onChange={(e) => this.setState({ selectedClass: e.value })} />
+                    <ToggleButton
+                      className={`relative-time-toggle-button ${relativeTime !== null ? "p-button-success" : "p-button-info"}`}
+                      checked={relativeTime === null}
+                      onLabel="Set"
+                      offLabel="Unset"
+                      onChange={() => {
+                        if (relativeTime === null) {
+                          this.setState({ relativeTime: selectedClass })
+                        } else {
+                          this.setState({ relativeTime: null })
+                        }
+                      }}
+                      style={{ borderRadius: "0 4px 4px 0" }}
+                    />
+                  </div>
+                </Col>
+              </Col>
+              <Col style={{ display: "flex", flexDirection: "column", justifyContent: "center", marginTop: "1rem" }}>
+                <label htmlFor="dd-city">Set time points to selected data points</label>
+                <Col style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginTop: "0rem" }}>
+                  <div style={{ width: "100%" }} className="p-inputgroup ">
+                    <Dropdown style={{ width: "100%" }} value={timePoint} options={timePoints} onChange={(e) => this.setState({ timePoint: e.value })} />
+                    <Button className="separate-toggle-button" style={{ borderRadius: "0 4px 4px 0", width: "5rem", padding: "0rem" }} onClick={this.handleSetTimePoint} label={`Set T${timePoint}`} />
+                  </div>
+                </Col>
+              </Col>
+              <Col style={{ display: "flex", flexDirection: "column", justifyContent: "center", marginTop: "1rem" }}>
+                <label htmlFor="dd-city">Set time points by classes</label>
+                <Col style={{ display: "flex", flexDirection: "row", justifyContent: "center", marginTop: "0rem" }}>
+                  <div style={{ width: "100%" }} className="p-inputgroup ">
+                    {/* <Dropdown style={{ width: "100%" }} value={selectedClassesToSetTimePoint} options={this.getClassesOptions()} onChange={(e) => this.setState({ selectedClassesToSetTimePoint: e.value })} /> */}
+                    <MultiSelect style={{ width: "100%" }} value={selectedClassesToSetTimePoint} options={this.getClassesOptions()} onChange={(e) => this.setState({ selectedClassesToSetTimePoint: e.value })} />
+                    <Button className="separate-toggle-button" style={{ borderRadius: "0 4px 4px 0", width: "5rem", padding: "0rem" }} onClick={this.handleSetTimePointByClass} label={`Set T${timePoint}`} />
+                  </div>
+                </Col>
+              </Col>
+              <Col style={{ display: "flex", flexDirection: "column", justifyContent: "center", marginTop: "1rem", alignItems: "flex-start", flex: "unset" }}>
+                {this.state.timePointClusters.length !== 0 && (
+                  <>
+                    <label htmlFor="">
+                      <b>
+                        <u>Time Points associated data</u>
+                      </b>
+                    </label>
+                    {this.state.timePointClusters.map((cluster, index) => {
+                      console.log("cluster", cluster)
+                      return (
+                        <>
+                          <div style={{ display: "flex", flexDirection: "row", alignContent: "center", alignItems: "flex-start", justifyContent: "center" }}>
+                            <h6 style={{ margin: "0" }}>{`T${cluster.name}`}</h6>
+                            &nbsp;
+                            <p style={{ margin: "0", marginLeft: "0.5rem" }}>
+                              {" "}
+                              {`Number of data points: `}
+                              <b> {`${cluster.x.length}`}</b>
+                            </p>
+                            <a
+                              value={cluster.name}
+                              style={{ margin: "0", marginLeft: "0.5rem", cursor: "pointer" }}
+                              onClick={() => {
+                                let newTimePointsClusters = deepCopy(timePointClusters)
+                                newTimePointsClusters.splice(index, 1)
+                                let newTimePoints = deepCopy(timePoints)
+                                let indexOfTimePoint = newTimePoints.findIndex((timePoint) => timePoint.value === cluster.name)
+                                console.log("indexOfTimePoint", indexOfTimePoint)
+                                if (indexOfTimePoint !== -1 && cluster.name !== 1) {
+                                  newTimePoints.splice(indexOfTimePoint, 1)
+                                }
+                                this.setState({ timePoints: newTimePoints })
+                                this.setState({ timePointClusters: newTimePointsClusters })
+                                this.removeTimePointFromJsonData(cluster.name)
+                              }}
+                            >
+                              <XSquare size={20} />
+                            </a>
+                          </div>
+                        </>
+                      )
+                    })}
+                  </>
+                )}
+              </Col>
+            </Row>
+          </Col>
+        </Row>
+      </>
+    )
+  }
 }
 
-export default MEDcohortFigure
+export default MEDcohortFigureClass
